Move realtime subscribers notification to separate module
